+++ src/Phan/Language/UnionType.php
@@ -28,8 +28,8 @@ use Phan\Language\Type\MultiType;
 use Phan\Language\Type\NullType;
 use Phan\Language\Type\ScalarType;
 use Phan\Language\Type\SelfType;
-use Phan\Language\Type\StaticType;
 use Phan\Language\Type\StaticOrSelfType;
+use Phan\Language\Type\StaticType;
 use Phan\Language\Type\StringType;
 use Phan\Language\Type\TemplateType;
 use Phan\Language\Type\TrueType;
@@ -3066,6 +3066,41 @@ class UnionType implements Serializable
             default:
                 return null;
         }
+    }
+
+    /**
+     * Returns true if this contains a type that is definitely nullable or a non-object.
+     * e.g. returns true for ?T, T|false, T|array
+     *      returns false for T|callable, object, T|iterable, etc.
+     */
+    public function containsDefiniteNonObjectType() : bool
+    {
+        foreach ($this->type_set as $type) {
+            if ($type->getIsNullable() || $type->isDefiniteNonObjectType()) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public function containsDefiniteNonCallableType() : bool
+    {
+        foreach ($this->type_set as $type) {
+            if ($type->getIsNullable() || $type->isDefiniteNonCallableType()) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public function hasPossiblyCallableType() : bool
+    {
+        foreach ($this->type_set as $type) {
+            if (!$type->isDefiniteNonCallableType()) {
+                return true;
+            }
+        }
+        return \count($this->type_set) === 0;
     }
 }