+++ NEWS.md
@@ -1,6 +1,27 @@
 Phan NEWS
 
-?? ??? 2018, Phan 1.1.2 (dev)
+?? ??? 2018, Phan 1.1.3 (dev)
+-----------------------
+
+New features (CLI)
++ Warn when calling method on union types that are definitely partially invalid. (#1885)
+  New config setting: `--strict-method-checking` (enabled as part of `--strict-type-checking`)
+  New issue type: `PhanPossiblyNonClassMethodCall`
++ Warn about invalid/possibly invalid callables in function calls.
+  New issue types: `PhanTypeInvalidCallable`, `PhanTypePossiblyInvalidCallable` (the latter check requires `--strict-method-checking`)
++ Add a prototype tool `tool/phoogle`, which can be used to search for function/method signatures in user-declared and internal functions/methods.
+  E.g. to look for functions that return a string, given a string and an array:
+  `/path/phan/tool/phoogle 'string -> array -> string`
+
+Plugins:
++ Add `BeforeAnalyzeCapability`, which will be executed once before starting the analysis phase. (#2086)
+
+New features (Analysis)
+
++ Add a heuristic check to detect potential infinite recursion in a functionlike calling itself (i.e. stack overflows)
+  New issue types: `PhanInfiniteRecursion`
+
+05 Nov 2018, Phan 1.1.2
 -----------------------
 
 New features(CLI)
@@ -36,6 +57,7 @@ New features(Analysis):
 + Properly warn about undefined namespaced constants/functions from within a namespace (#2112)
   Phan was failing to warn in some cases.
 + Always infer `int` for `<<` and `>>`
++ Support using dynamic values as the name for a `define()` statement (#2116)
 
 Maintenance:
 + Make issue messages more consistent in their syntax used to describe closures/functions (#1695)